+++ src/lib/onboard.ts
@@ -12,17 +12,17 @@ export async function prompt_configure_extensions(
 	executable: string
 ): Promise<string> {
 	const selection_map: Record<string, () => Promise<string>> = {
-		'Always install extensions (recommended)': async () => {
+		'Always use extensions (recommended)': async () => {
 			await set_config_exclusive('renpyExtensionsEnabled', 'Enabled')
 
 			return 'Enabled'
 		},
-		'Use extensions in this project': async () => {
+		'Use extensions only in this project': async () => {
 			await set_config('renpyExtensionsEnabled', 'Enabled', true)
 
 			return 'Enabled'
 		},
-		'Do not use extensions in this project': async () => {
+		'Disable extensions only in this project': async () => {
 			await set_config('renpyExtensionsEnabled', 'Disabled', true)
 
 			vscode.window.showInformationMessage(